Optimize memory layout of the UI classes xrUICore/uiabstract.h: default initialize members of CUISimpleWindow

+++ src/xrUICore/ScrollBar/UIScrollBar.h
@@ -30,7 +30,7 @@ protected:
     bool m_b_enabled{ true };
     bool m_bIsHorizontal{ false };
 
-    int m_mouse_state{};
+    u8 m_mouse_state{};
 
     bool ScrollInc(bool by_scrollbox = false);
     bool ScrollDec(bool by_scrollbox = false);
